corrigiendo errores repositories

+++ src/main/java/com/veterinary/practica/domains/repositories/clients/SocialNetworksClientsRepository.java
@@ -3,5 +3,4 @@ package com.veterinary.practica.domains.repositories.clients;
 import com.veterinary.practica.domains.entities.clients.SocialNetworksClientsEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SocialNetworksClientsRepository extends CrudRepository<SocialNetworksClientsEntity, Long> {
-}
+public interface SocialNetworksClientsRepository extends CrudRepository<SocialNetworksClientsEntity, Integer> {}
